Refactor serviceWorker and api.ts
This commit refactors the serviceWorker.ts and api.ts files. In serviceWorker.ts, a new event listener has been added to handle the "showOptions" message. When this message is received, the options page is opened and an alert is displayed to inform the user that the API Key is not set. In api.ts, the alert message for the API Key not being set has been commented out and replaced with a call to the chrome.runtime.sendMessage() function to send the "showOptions" message.